Added missing WS plugin and fixed warnings

+++ objective-c/Chat/Cocoa/Classes/ChatController.m
@@ -263,7 +263,7 @@
 
 - (int)numberOfRowsInTableView:(NSTableView *)tv
 {
-    return users.count;
+    return (int)users.count;
 }
 
 -(id) tableView:(NSTableView *)tv
